changes the apis db connect import

+++ src/app/(dashboard)/dashboard/companies/api/delete-company/[id]/route.ts
@@ -1,4 +1,4 @@
-import { connectToDatabase } from "@/app/api/auth/[...nextauth]/route";
+import { connectDB } from "@/lib/ConnectDB";
 import { ObjectId } from "mongodb";
 import { NextResponse } from "next/server";
 
@@ -6,12 +6,14 @@ export const DELETE = async (
   req: Request,
   { params }: { params: { id: string } }
 ) => {
-  const db = await connectToDatabase(); // Ensure the database connection is awaited
+  const db = await connectDB(); // Ensure the database connection is awaited
 
   try {
     // Ensure the id is a valid ObjectId
     const objectId = new ObjectId(params.id);
-
+    if (!db) {
+      throw new Error("Failed to connect to the database");
+    }
     // Delete the document with the matching _id
     const res = await db.collection("companies").deleteOne({ _id: objectId });
